requirements.txt
weeve_modules
python-snap7
